API for Dark
